#96 [feat] 코드리뷰 반영
# Conflicts:
#	src/main/java/org/sopt/certi_server/domain/certification/service/CertificationService.java
